Build the package, check in the result, prepare for npm publish
Summary: This commit prepares this package for actually being used in webapp
and on npm. It defines our build script, and checks in the result (because we
need that in webapp). Also, I added an `.npmignore` and set the main file to
the built version (do we still do that? it's 2017!) so that we can prepare for
publishing this to npm.

Test Plan:
 - Run `npm pack`
 - Run `tar -tvf declarative-z-indexes-1.0.0.tgz`, see that all of the files
   that I want are in the list of files to be published.
 - Try using the built `lib/index.js` file in webapp, and see that it works.

Auditors: joshua